Separate backend jobs
Motivation

The actual backend job is responsible for fetching data from the football API and to transform data to compute aggregate stats and others metrics. Then the data is stored into MongoDB.
The first part has some complexity because the football API have rate limits so we can't call it too often. But once the data is in MongoDB, we can transform the data as many times as we want.
So the objective is to split this job into two different jobs. The first job will continue to fetch data from the football and store it into a new collection players_football_api_<season>.
We added the season to be able to support multiple year in the future.

- We added a status on the job type. This is better for monitoring and helps to have the job status when the rate limit is reached for the day.
- Instead of returning an error when the rate limit is reached, we stop the current job properly.
- New season type