Fix players search (#30)
## Dev notes

There are some issues in the way the search is handled in the frontend part. We use the reactivity statement feature from Svelte.
But the reactive block is triggered too many times because the props of the component are not initialized. Then each time the prop receives its value the reactive block is triggered.
In order to avoid this, we added some logic to prevent an API call but this leads to some bugs. When filters are reset, no updates were performed.
A workaround is to use the onMount function to wait for the component to be mounted. The result is the reactive block is triggered only once when the component is mounted. This is not ideal but it fixes the filters issue. In a perfect world, we do not want to make an API call at all.